+++ packages/tests/tests/light-push/index.node.spec.ts
@@ -74,24 +74,13 @@ describe("Waku Light Push", function () {
     }
   });
 
-  it("Fails to push message with empty payload", async function () {
+  it("Throws when trying to push message with empty payload", async function () {
     const pushResponse = await waku.lightPush.send(TestEncoder, {
-      payload: utf8ToBytes("")
+      payload: new Uint8Array()
     });
 
-    if (nwaku.type() == "go-waku") {
-      expect(pushResponse.recipients.length).to.eq(1);
-      expect(await messageCollector.waitForMessages(1)).to.eq(true);
-      messageCollector.verifyReceivedMessage(0, {
-        expectedMessageText: undefined,
-        expectedContentTopic: TestContentTopic
-      });
-    } else {
-      expect(pushResponse.recipients.length).to.eq(0);
-      // This should be `REMOTE_PEER_REJECTED`, tracked with https://github.com/waku-org/nwaku/issues/1641
-      expect(pushResponse.errors).to.include(SendError.REMOTE_PEER_FAULT);
-      expect(await messageCollector.waitForMessages(1)).to.eq(false);
-    }
+    expect(pushResponse.errors).to.include(SendError.EMPTY_PAYLOAD);
+    expect(await messageCollector.waitForMessages(1)).to.eq(false);
   });
 
   TEST_STRING.forEach((testItem) => {